add database in setting with float

+++ blog/forms.py
@@ -9,10 +9,10 @@ class PostForm(forms.ModelForm):
         fields = ('title', 'text',)
 
 class dtForm(forms.ModelForm):
-    """pH=forms.IntegerField()
-    temp=forms.IntegerField()
-    Ec=forms.IntegerField()
-    Water=forms.IntegerField()"""
+    """pH=forms.DecimalField()
+    temp=forms.DecimalField()
+    Ec=forms.DecimalField()
+    Water=forms.DecimalField()"""
     class Meta:
         model = dt
         fields = (
@@ -22,10 +22,15 @@ class dtForm(forms.ModelForm):
             'Water',
         )
 class dtForm2(forms.ModelForm):
-    """pH=forms.IntegerField()
-    temp=forms.IntegerField()
-    Ec=forms.IntegerField()
-    Water=forms.IntegerField()"""
+    """pH=forms.DecimalField()
+    temp=forms.DecimalField()
+    Ec=forms.DecimalField()
+    Water=forms.DecimalField()"""
+    # pH=forms.DecimalField(..., max_digits=1, decimal_places=2)
+    # pH=forms.IntegerField(required=True)
+    # temp=forms.IntegerField(required=False)
+    # Ec=forms.IntegerField(required=False)
+    # Water=forms.IntegerField(required=False)
     class Meta:
         model = dt2
         fields = (
